Updated Silver's Extra Dusty Desert, slightly updated Omega's Extra aquatic base, Updated Tails White Acropolis,added ring explosion patch.(read description)
To all those who have supported this mod, thank you! For the time being, after many years of updates to get this mod at the state I think it deserves to be at, and finally achieving that goal. I am pleased to announce that I am about done with updates, I am looking into onre more stage idea with Amy, but that will take some time. In the meantime, please note unless it's a bugfix, there will not be any LoS updates for the forseeable future.